Move bridge to its own package
